requirements.txt
alembic>=0.8.8
altgraph>=0.10.2
aniso8601>=1.3.0
bdist-mpkg>=0.5.0
certifi>=2017.7.27.1
cffi>=1.10.0
chardet>=3.0.4
click>=6.7
cryptography>=1.7.2
enum34>=1.1.6
Flask-Cors>=3.0.3
Flask-JWT-Extended>=3.3.4
Flask-Migrate>=2.0.0
Flask-Moment>=0.5.1
Flask-RESTful>=0.3.6
Flask-Script>=2.0.5
Flask-SQLAlchemy>=2.1
Flask-WTF>=0.13.1
idna>=2.5
psycopg2>=2.8.3
psycopg2-binary>=2.8.3
ipaddress>=1.0.18
itsdangerous>=0.24
macholib>=1.5.1
Mako>=1.0.5
MarkupSafe>=1.0
matplotlib>=1.3.1
modulegraph>=0.10.4
Moments>=2.0
numpy>=1.8.0rc1
passlib>=1.7.1
py2app>=0.7.3
pyasn1>=0.2.3
pycparser>=2.17
PyJWT>=1.5.3
pyparsing>=2.0.1
python-dateutil>=1.5
python-editor>=1.0.1
pytz>=2013.7
scipy>=0.13.0b1
six>=1.11.0
virtualenv>=15.0.2
pyenv>=1.2.13
Werkzeug>=0.12.2
WTForms>=2.1
xattr>=0.6.4
zope.interface>=4.1.1
SQLAlchemy>=1.3.0
Jinja2>=2.10.1
flask>=0.12.3
urllib3>=1.24.2
requests>=2.20.0
pyopenssl>=17.5.0
